requirement.txt
pynput.keyboard
smtplib